back/requirements.txt
alembic==1.13.3
anyio==4.10.0
bcrypt==4.3.0
blinker==1.9.0
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.8.3
cffi==2.0.0
click==8.2.1
cssselect2==0.8.0
defusedxml==0.7.1
Flask==3.0.3
flask-cors==6.0.1
Flask-JWT-Extended==4.6.0
greenlet==3.0.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
playwright==1.47.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
pycparser==2.23
pyee==12.0.0
PyJWT==2.10.1
python-docx==1.2.0
python-dotenv==1.0.1
redis==5.3.1
rq==1.16.2
sniffio==1.3.1
SQLAlchemy==2.0.43
tinycss2==1.4.0
typing_extensions==4.15.0
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
